refactor all, structure, make pretty
