+++ flixkit/fcl_bindings.go
@@ -1,14 +1,83 @@
-package bindings
+package flixkit
 
 import (
-	"net/url"
+	"bytes"
+	"fmt"
 	"sort"
 	"text/template"
 
-	"github.com/onflow/flixkit-go"
-	v1_1 "github.com/onflow/flixkit-go/flixkitv1_1"
+	"github.com/onflow/flixkit-go/flixkit/v1"
+	v1_1 "github.com/onflow/flixkit-go/flixkit/v1_1"
+	"github.com/onflow/flixkit-go/internal/templates"
 	"github.com/stoewer/go-strcase"
 )
+
+func NewFclTSGenerator() *FclGenerator {
+	t := []string{
+		templates.GetTsFclMainTemplate(),
+		templates.GetTsFclScriptTemplate(),
+		templates.GetTsFclTxTemplate(),
+		templates.GetTsFclParamsTemplate(),
+		templates.GetTsFclInterfaceTemplate(),
+	}
+
+	return &FclGenerator{
+		Templates: t,
+	}
+}
+
+func NewFclJSGenerator() *FclGenerator {
+	t := []string{
+		templates.GetJsFclMainTemplate(),
+		templates.GetJsFclScriptTemplate(),
+		templates.GetJsFclTxTemplate(),
+		templates.GetJsFclParamsTemplate(),
+	}
+
+	return &FclGenerator{
+		Templates: t,
+	}
+}
+
+func (g *FclGenerator) Generate(flixString string, templateLocation string) (string, error) {
+	tmpl, err := parseTemplates(g.Templates)
+	if err != nil {
+		return "", err
+	}
+	if flixString == "" {
+		return "", fmt.Errorf("no flix template provided")
+	}
+	isLocal := !isUrl(templateLocation)
+
+	ver, err := GetTemplateVersion(flixString)
+	if err != nil {
+		return "", fmt.Errorf("invalid flix template version, %s", err)
+	}
+	var data templateData
+	data.FclVersion = GetFlixFclCompatibility(ver)
+	if ver == "1.0.0" {
+		flix, err := flixkit.ParseFlix(flixString)
+		if err != nil {
+			return "", err
+		}
+		data = getTemplateDataV1_0(flix, templateLocation, isLocal)
+
+	} else if ver == "1.1.0" {
+		flix, err := v1_1.ParseFlix(flixString)
+		if err != nil {
+			return "", err
+		}
+		data = getTemplateDataV1_1(flix, templateLocation, isLocal)
+
+	} else {
+		return "", fmt.Errorf("invalid flix template version, support v1.0.0 and v1.1.0")
+	}
+	data.FclVersion = GetFlixFclCompatibility(ver)
+
+	var buf bytes.Buffer
+	err = tmpl.Execute(&buf, data)
+	return buf.String(), err
+}
 
 type simpleParameter struct {
 	Name        string
@@ -38,11 +107,6 @@ type FclGenerator struct {
 type FlixParameter struct {
 	Name string
 	Type string
-}
-
-func isUrl(str string) bool {
-	u, err := url.Parse(str)
-	return err == nil && u.Scheme != "" && u.Host != ""
 }
 
 func GetFlixFclCompatibility(flixVersion string) string {
