+++ flixkit/testdata/TestJSGenArrayScript.golden
