+++ flixkit/testdata/TestJSGenMinScript.golden
