+++ flixkit/testdata/TestJSGenNoParamsScript.golden
