+++ internal/templates/ts_fcl_main.tmpl.go
@@ -1,4 +1,4 @@
-package bindings
+package templates
 
 func GetTsFclMainTemplate() string {
 	const template = `/**
