+++ internal/templates/ts_fcl_tx.tmpl.go
@@ -1,4 +1,4 @@
-package bindings
+package templates
 
 func GetTsFclTxTemplate() string {
 	const template = `{{define "tx"}}export async function {{.Title}}(
